+++ Src/Private/Get-AbrAppVolADDomain.ps1
@@ -5,7 +5,7 @@ function Get-AbrAppVolADDomain {
     .DESCRIPTION
         Documents the configuration of VMware APPVolume in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.2.0
+        Version:        1.1.0
         Author:         Chris Hildebrandt, @childebrandt42
         Editor:         Jonathan Colon, @jcolonfzenpr
         Twitter:        @asbuiltreport
@@ -34,28 +34,36 @@ function Get-AbrAppVolADDomain {
                 } else {$LDAPDomains = Invoke-RestMethod -WebSession $SourceServerSession -Method Get -Uri "https://$AppVolServer/cv_api/ldap_domains"}
                 if ($LDAPDomains) {
                     section -Style Heading3 "Active Directory Domain" {
+                        Paragraph "The following section details active directory doamins are used for authentication on $($AppVolServer.split('.')[0])."
+                        BlankLine
                         $OutObj = @()
                         foreach ($LDAPDomain in $LDAPDomains.ldap_domains | Sort-Object -Property Domain) {
-                            section -Style Heading3 $LDAPDomain.domain {
+
+                            If($LDAPDomain.ldaps -like 'True'){
+                                $Security = 'LADPS'
+                            }elseif($LDAPDomain.ldaps -like 'False' -and $LDAPDomain.ldap_tls -like 'False'){
+                                $Security = 'LADP'
+                            }
+                            elseif($LDAPDomain.ldap_tls -like 'True'){
+                                $Security = 'LADPS over TLS'
+                            }
+                            section -Style Heading4 "AD Domain Summary"  {
                                 try {
                                     $inObj = [ordered] @{
-                                        'Username' = $LDAPDomain.username
-                                        'Base' = $LDAPDomain.base
+                                        'Domain' = $LDAPDomain.domain
                                         'NetBIOS' = $LDAPDomain.netbios
-                                        'LDAPS' = $LDAPDomain.ldaps
-                                        'LDAP_TLS' = $LDAPDomain.ldap_tls
+                                        'Base' = $LDAPDomain.base
+                                        'Username' = $LDAPDomain.username
+                                        'Security' = $Security
                                         'SSL Verify' = $LDAPDomain.ssl_verify
-                                        'Port' = $LDAPDomain.port
-                                        'Effective Port' = $LDAPDomain.effective_port
-                                        'Created At' = $LDAPDomain.created_at
-                                        'Updated At' = $LDAPDomain.updated_at
+                                        'Port' = $LDAPDomain.effective_port
                                     }
-                                    $OutObj = [pscustomobject](ConvertTo-HashToYN $inObj)
+                                    $OutObj += [pscustomobject](ConvertTo-HashToYN $inObj)
 
                                     $TableParams = @{
-                                        Name = "Active Directory Domains - $($LDAPDomain.domain)"
-                                        List = $true
-                                        ColumnWidths = 50, 50
+                                        Name = "AD Domain Summary - $($AppVolServer)"
+                                        List = $false
+                                        ColumnWidths = 20, 20, 15, 15, 10, 10, 10
                                     }
                                     if ($Report.ShowTableCaptions) {
                                         $TableParams['Caption'] = "- $($TableParams.Name)"
@@ -64,6 +72,43 @@ function Get-AbrAppVolADDomain {
                                 }
                                 catch {
                                     Write-PscriboMessage -IsWarning $_.Exception.Message
+                                }
+                                if ($InfoLevel.AppVolumes.ADDomains -ge 2) {
+                                    $OutObj = @()
+                                    foreach ($LDAPDomain in $LDAPDomains.ldap_domains | Sort-Object -Property Domain) {
+                                        section -ExcludeFromTOC -Style NOTOCHeading5 "AD Domain Details - $($LDAPDomain.domain)" {
+                                            try {
+                                                $inObj = [ordered] @{
+                                                    'Username' = $LDAPDomain.username
+                                                    'Base' = $LDAPDomain.base
+                                                    'NetBIOS' = $LDAPDomain.netbios
+                                                    'LDAPS' = $LDAPDomain.ldaps
+                                                    'LDAP_TLS' = $LDAPDomain.ldap_tls
+                                                    'SSL Verify' = $LDAPDomain.ssl_verify
+                                                    'Port' = $LDAPDomain.effective_port
+                                                    'Created At' = $LDAPDomain.created_at
+                                                    'Updated At' = $LDAPDomain.updated_at
+                                                            }
+                                                $OutObj += [pscustomobject](ConvertTo-HashToYN $inObj)
+
+
+                                                $TableParams = @{
+                                                    Name = "AD Domain Details - $($LDAPDomain.domain)"
+                                                    List = $true
+                                                    ColumnWidths = 50, 50
+                                                }
+                                                if ($Report.ShowTableCaptions) {
+                                                    $TableParams['Caption'] = "- $($TableParams.Name)"
+                                                }
+                                                $OutObj | Table @TableParams
+                                            }
+                                            catch {
+                                                Write-PscriboMessage -IsWarning $_.Exception.Message
+                                            }
+                                        }
+                                    }
+
+
                                 }
                             }
                         }